+++ src/core/queue/worker/simple/CHANGELOG.md
@@ -9,6 +9,13 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.? (2022-0?-??)
+
+#### :rocket: New Feature
+
+* Implemented a new `clone` method
+* Implemented iterable behavior
+
 ## v3.22.2 (2020-07-21)
 
 #### :house: Internal